[DELETE] : SearchViewmodel viewmodel 로직 삭제 #54
